+++ R/zzz.R
@@ -3,6 +3,7 @@
   register_s3_method("dbplyr", "db_compute", "AthenaConnection")
   register_s3_method("dplyr", "db_save_query", "AthenaConnection")
   register_s3_method("dbplyr", "db_copy_to", "AthenaConnection")
+  register_s3_method("dbplyr", "sql_translate_env", "AthenaConnection")
 }
 
 register_s3_method <- function(pkg, generic, class, fun = NULL) {